+++ src/script/Context.cpp
@@ -227,7 +227,7 @@ Context::Context(){
   _globals[Symbol(#op)] = (Native)([](Context& c,const Array<Var>& a)->Var {\
     L_ASSERT(a.size()>1);\
     Var* target(c.reference(a[1]));\
-    for(int i(2);i<a.size();i++)\
+    for(uintptr_t i(2);i<a.size();i++)\
       *target op c.execute(a[i]);\
     return 0;\
   })
